requirements.txt
streamlit
streamlit-chat
python-socketio
eventlet
pymongo
openai
python-dotenv